extract shared RichTextRenderer and fix FAQ answer rendering
